==> 0. Data_preparing.ipynb <==
import warnings
import os
import pandas as pd
import numpy as np
import tqdm

import sys

sys.path.insert(1, '/Users/gosha/Desktop/DS Competitions/MTS Cup/src')

warnings.filterwarnings('ignore')

# импортируем свои функции для обработки данных
from src.preprocessing import get_data_part_day, get_data_days, 
    get_user_model_price, get_user_url_cnt, get_user_city_cnt

# Описание задачи

**В данном jupyter-ноутбуке мы рассмотрим и предобработаем 1 из 10 файлов с данными. После чего аггрегируем отдельно каждый файл в цикле и соединим в 1 датафрейм**

## О соревновании и данных

https://ods.ai/competitions/mtsmlcup

Задача соревнования
- Определение пола и возраста владельца HTTP cookie по истории активности пользователя в интернете на основе синтетических данных.

Метрики соревнования:
* ROC-AUC – для определения пола, f1 weighted – для определения возраста.
* Все решения рассчитываются по формуле -  2 * f1_weighted(по 6 возрастным бакетам) + gini по полу.
* Возрастные бакеты (Класс 1 — 19-25, Класс 2 — 26-35, Класс 3 — 36-45, Класс 4 — 46-55, Класс 5 — 56-65, Класс 6 — 66+).

Описание колонок файла с данными:
* 'region_name' – Регион
* 'city_name' – Населенный пункт
* 'cpe_manufacturer_name' – Производитель устройства
* 'cpe_model_name' – Модель устройства
* 'url_host' – Домен, с которого пришел рекламный запрос
* 'cpe_type_cd' – Тип устройства (смартфон или что-то другое)
* 'Cpe_model_os_type' – Операционка на устройстве
* 'price' – Оценка цены устройства
* 'date' – Дата
* 'part_of_day' – Время дня (утро, вечер, и тд)
* 'request_cnt' – Число запросов одного пользователя за время дня (поле part_of_day)
* 'user_id' – ID пользователя

Описание колонок файла с таргетами:

* 'age' – Возраст пользователя
* 'is_male' – Признак пользователя : мужчина (1-Да, 0-Нет)
* 'user_id' – ID пользователя

## Структура проекта

Данная работа была разделена на несколько jupyter ноутбуков:

0. Data_preparing.ipnb - аггрегация отдельных файлов по user_id и склейка в финальный датасет
1. EDA.ipynb - исследовательская часть
2. baseline.ipynb - бейзлайн модели
3. create_embeddings.ipynb - создание эмбеддингов для дальнейшего их использования в качестве фич
4. baseline_embeddings.ipynb - бейзлан модели с эмбеддингами
5. model_tuning.ipynb - тюнинг наиболее перспективных моделей
6. gender_prediction_stacking.ipynb - стекинг моделей для предсказания пола

# Data preparing

df = pd.read_parquet(
    '../competition_data_final_pqt/part-00000-aba60f69-2b63-4cc1-95ca-542598094698-c000.snappy.parquet',
    engine='fastparquet')

df.shape

df.head(3)

# поменяем тип числовых признаков, чтобы сэкономить память
df = df.astype({'request_cnt': 'int8', 'price': 'float32', 'user_id': 'int32'})

df.info()

df.isna().sum()

def replace_model_mistakes(data: pd.DataFrame) -> pd.DataFrame:
    """
    Функция исправляет неточности в данных, заменяет
    :param data: датафрейм с данными
    :return: датафрейм с исправленными неточностями
    """
    # заменим значения, которые являются одинаковыми, но записаны по-разному
    data.cpe_model_os_type.replace('Apple iOS', 'iOS', inplace=True)
    data.cpe_manufacturer_name.replace('Huawei Device Company Limited',
                                       'Huawei',
                                       inplace=True)
    data.cpe_manufacturer_name.replace(
        'Motorola Mobility LLC, a Lenovo Company', 'Motorola', inplace=True)
    data.cpe_manufacturer_name.replace('Sony Mobile Communications Inc.',
                                       'Sony',
                                       inplace=True)

    # в таблицах для Nokia 3 Dual попадаются разные типы устройства, оставим только один тип
    data['cpe_type_cd'] = np.where((data['cpe_manufacturer_name'] == 'Nokia') &
                                   (data['cpe_model_name'] == '3 Dual'),
                                   'plain', data['cpe_type_cd'])

    # заполним пропуски цены несуществующим значением
    data.price.fillna(-999, inplace=True)

    return data

df = replace_model_mistakes(df)

# Feature Generation

## Количество запросов в разные части дня

"""
Аггрегируем датафрейм по user_id и генерируем новые признаки. Итоговый датафрейм содержит признаки:
- part_of_day_day - кол-во визитов в часть дня day
- part_of_day_evening - кол-во визитов в часть дня evening
- part_of_day_morning - кол-во визитов в часть дня morning
- part_of_day_night - кол-во визитов в часть дня night
- sum_visits - общее кол-во визитов
- day_pct - доля визитов в часть дня day
- evening_pct - доля визитов в часть дня evening
- morning_pct - доля визитов в часть дня morning
- night_pct - доля визитов в часть дня night
"""
df_part_day = get_data_part_day(df)
df_part_day.head(3)

## Количество дней активности, среднее кол-во запросов в сутки

"""
Аггрегируем датафрейм по user_id и генерируем новые признаки. Итоговый датафрейм содержит признаки:
- act_days - кол-во дней, в которые пользователь совершил визит пользователя
- request_cnt - кол-во запросов пользователя
- avg_req_per_day - среднее кол-во запросов пользователя
- period_days - кол-во дней между первым и последним визитом пользователя
- request_std - стандартное отклонение по количеству запросов
- act_days_pct - доля дней, когда пользователь совершал визит
"""
df_days = get_data_days(df)
df_days.head(3)

## Цена устройства

"""
Аггрегируем датафрейм по user_id и генерируем новые признаки. Итоговый датафрейм содержит признаки:
- cpe_type_cd - тип устройства
- cpe_model_os_type - операционная система устройства
- cpe_manufacturer_name - производитель устройства
- price - цена устройства пользователя
"""
df_model = get_user_model_price(df)
df_model.head(3)

## Регион и город

"""
Аггрегируем датафрейм по user_id и генерируем новые признаки. Итоговый датафрейм содержит признаки:
- region_cnt - кол-во уникальных регионов, из которых был совершен визит
- city_cnt - кол-во уникальных городов, из которых был совершен визит
"""
df_city_cnt = get_user_city_cnt(df)
df_city_cnt.head(3)

## URLs

"""
Аггрегируем датафрейм по user_id и генерируем новые признаки. Итоговый датафрейм содержит признаки:
- url_host_cnt - кол-во уникальных ссылок, с которых был совершен визит
"""
df_url = get_user_url_cnt(df)
df_url.head(3)

# Итоговый датафрейм

Теперь проделаем те же действия для каждого файла с данными и объединим все в общий датафрейм

# название папки с файлами
DATA_FOLD = '../competition_data_final_pqt'

df_final = pd.DataFrame()
for file in tqdm.tqdm_notebook(os.listdir(DATA_FOLD)):
    if file.endswith('.parquet'):
        temp_df = pd.read_parquet(f'{DATA_FOLD}/{file}', engine='fastparquet')
        temp_df = replace_model_mistakes(temp_df)

        # получаем аггрегированные данные
        # кол-во визитов пользователя в разное время суток
        data_part_day = get_data_part_day(temp_df)
        # кол-во активных дней и среднее кол-во запросов в сутки
        data_days = get_data_days(temp_df)
        # данные об устройстве
        data_user_model = get_user_model_price(temp_df)
        # кол-во регионов и городов
        data_city_cnt = get_user_city_cnt(temp_df)
        # кол-во уникальных url
        data_url_cnt = get_user_url_cnt(temp_df)

        # объединяем все в один датафрейм
        temp_df_agg = (data_part_day
                       .merge(data_days, how='left', on='user_id')
                       .merge(data_user_model, how='left', on='user_id')
                       .merge(data_city_cnt, how='left', on='user_id')
                       .merge(data_url_cnt, how='left', on='user_id'))

        # добавляем аггрегированные данные в итоговый датафрейм
        df_final = pd.concat([df_final, temp_df_agg])

df_final.head()

df_final.shape

# проверим, что файлы не пересекались по user_id
(df_final.groupby('user_id', as_index=False)
         .agg({'city_cnt': 'count'})
         .query('city_cnt > 1'))

# сохраняем итоговый аггрегированный датафрейм
df_final.to_csv('../data_agg/df_final.csv', index=False)
