==> 6. gender_prediction_stacking.ipynb <==
# Импорт библиотек

from google.colab import drive
drive.mount('/content/drive')

%cd /content/drive/MyDrive/MTS ML Cup

import warnings
import os
import sys

sys.path.insert(0, '/content/drive/MyDrive/MTS ML Cup/src')

os.environ['OPENBLAS_NUM_THREADS'] = '1'
warnings.filterwarnings('ignore')

!pip install catboost

!pip install lightgbm

import pandas as pd
import numpy as np

from catboost import CatBoostClassifier
from lightgbm import LGBMClassifier

import sklearn.metrics as m
from sklearn.model_selection import train_test_split, StratifiedKFold
from sklearn.linear_model import LogisticRegression

from metric_functions import get_metrics_classification, check_overfitting_classification

import seaborn as sns
import matplotlib.pyplot as plt

%matplotlib inline
sns.set_style('darkgrid')

RAND = 42
N_FOLDS = 5

final_types = {
    'user_id': 'int32',
    'part_of_day_day': 'int16',
    'part_of_day_night': 'int16',
    'day_pct': 'float32',
    'evening_pct': 'float32',
    'morning_pct': 'float32',
    'night_pct': 'float32',
    'act_days': 'int16',
    'avg_req_per_day': 'float32',
    'requests_std': 'float32',
    'act_days_pct': 'float32',
    'cpe_type_cd': 'category',
    'cpe_model_os_type': 'category',
    'cpe_manufacturer_name': 'category',
    'price': 'float32',
    'region_cnt': 'int8',
    'city_cnt': 'int8',
    'url_host_cnt': 'int16',
}

# Описание задачи

**В данном jupyter-ноутбуке мы сделаем стекинг для предсказания пола из моделей с параметрами, подобранными в model_tuning.ipnb. Для стекинга будем использовать 3 модели LightGBM и 1 модель CatBoost**

## О соревновании и данных

https://ods.ai/competitions/mtsmlcup

Задача соревнования
- Определение пола и возраста владельца HTTP cookie по истории активности пользователя в интернете на основе синтетических данных.

Метрики соревнования:
* ROC-AUC – для определения пола, f1 weighted – для определения возраста.
* Все решения рассчитываются по формуле -  2 * f1_weighted(по 6 возрастным бакетам) + gini по полу.
* Возрастные бакеты (Класс 1 — 19-25, Класс 2 — 26-35, Класс 3 — 36-45, Класс 4 — 46-55, Класс 5 — 56-65, Класс 6 — 66+).

Предобработка, аггрегация и создание новых фич произведена в 0.Data_preparing.ipng.
Описание колонок аггрегированного файла с данными:
* 'part_of_day_day' – кол-во визитов пользователя днем
* 'part_of_day_evening' – кол-во визитов пользователя вечером
* 'part_of_day_morning' – кол-во визитов пользователя утром
* 'part_of_day_night' – кол-во визитов пользователя ночью
* 'sum_visits' – кол-во визитов пользователя
* 'day_pct' – доля визитов пользователя днем
* 'evening_pct' – доля визитов пользователя вечером
* 'morning_pct' – доля визитов пользователя утром
* 'night_pct' – доля визитов пользователя ночью
* 'act_days' – кол-во дней, в которые пользователь совершил визит пользователя
* 'request_cnt' - кол-во запросов пользователя
* 'avg_req_per_day' - среднее кол-во запросов пользователя
* 'period_days' - кол-во дней между первым и последним визитом пользователя
* 'request_std' - стандартное отклонение по количеству запросов
* 'act_days_pct' - доля дней, когда пользователь совершал визит
* 'cpe_type_cd - тип устройства
* 'cpe_model_os_type' - операционная система устройства
* 'cpe_manufacturer_name' -производитель устройства
* 'price' - цена устройства пользователя
* 'region_cnt' - кол-во уникальных регионов, из которых был совершен визит
* 'city_cnt' - кол-во уникальных городов, из которых был совершен визит
* 'url_host_cnt' - кол-во уникальных ссылок, с которых был совершен визит
* 'user_id' – ID пользователя

* также сгенерированы 300 признаков - эмбеддингов на основании региона, города, url и модели телефона

Описание колонок файла с таргетами:

* 'age' – Возраст пользователя
* 'is_male' – Признак пользователя : мужчина (1-Да, 0-Нет)
* 'user_id' – ID пользователя

## Структура проекта

Данная работа была разделена на несколько jupyter ноутбуков:

0. Data_preparing.ipnb - аггрегация отдельных файлов по user_id и склейка в финальный датасет
1. EDA.ipynb - исследовательская часть
2. baseline.ipynb - бейзлайн модели
3. create_embeddings.ipynb - создание эмбеддингов для дальнейшего их использования в качестве фич
4. baseline_embeddings.ipynb - бейзлан модели с эмбеддингами
5. model_tuning.ipynb - тюнинг наиболее перспективных моделей
6. gender_prediction_stacking.ipynb - стекинг моделей для предсказания пола


# Сбор всех данных

# id пользователей, которым нужно предсказать пол и возраст
id_to_submit = pd.read_parquet('submit.pqt')

# таргеты для обучения моделей
targets = pd.read_parquet('public_train.pqt')

# предобработанные нами данные
df = pd.read_csv('data_agg/df_final.csv')

df.head()

df = df.astype(final_types)
df.dtypes

df.shape

# загружаем созданные по url эмбеддинги
url_emb = pd.read_csv('embeddings/url_emb.csv')
url_emb = url_emb.add_suffix('_url').rename(columns={'user_id_url': 'user_id'})

# загружаем созданные по регионам эмбеддинги
reg_emb = pd.read_csv('embeddings/reg_emb.csv')
reg_emb = reg_emb.add_suffix('_reg').rename(columns={'user_id_reg': 'user_id'})

# загружаем созданные по городам эмбеддинги
city_emb = pd.read_csv('embeddings/city_emb.csv')
city_emb = city_emb.add_suffix('_city').rename(
    columns={'user_id_city': 'user_id'})

# загружаем созданные по девайсам эмбеддинги
model_emb = pd.read_csv('embeddings/model_emb.csv')
model_emb = model_emb.add_suffix('_model').rename(
    columns={'user_id_model': 'user_id'})

# Split data

# создаем датасет для обучения моделей
df_train_cat = targets.merge(df, how='inner', on=['user_id']) \
                      .merge(url_emb, how='inner', on='user_id') \
                      .merge(city_emb, how='inner', on=['user_id']) \
                      .merge(reg_emb, how='inner', on=['user_id']) \
                      .merge(model_emb, how='inner', on=['user_id'])

# удаляем пропуски
df_train_cat = df_train_cat[df_train_cat['is_male'] != 'NA']
df_train_cat = df_train_cat.dropna()

df_train_cat['is_male'] = df_train_cat['is_male'].map(int)
df_train_cat['is_male'].value_counts()

# сплит данных на train/test
%%time
X = df_train_cat.drop(['user_id', 'age', 'is_male'], axis=1)
y = df_train_cat['is_male']

X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    test_size=0.2,
                                                    shuffle=True,
                                                    stratify=y,
                                                    random_state=RAND)

cat_features = X.select_dtypes('category').columns.tolist()
# значение для scale_pos_weight
ratio = float(np.sum(y == 0)) / np.sum(y == 1)

# Metrics

# параметры для моделей были подобраны в model_tuning.ipnb с помощью Optuna
lgbm_gender_best1 = {
    'n_estimators': 1600,
    'learning_rate': 0.04323025519185361,
    'num_leaves': 900,
    'max_depth': 7,
    'min_child_samples': 1200,
    'reg_alpha': 0,
    'reg_lambda': 9,
    'min_split_gain': 3,
    'subsample': 0.6909275901887844,
    'subsample_freq': 1,
    'colsample_bytree': 0.34716278523119637,
    'scale_pos_weight': 0.9531740695036613,
    'random_state': 42
}

lgbm_gender_params_2 = {
    'n_estimators': 2200,
    'learning_rate': 0.04667738692602193,
    'scale_pos_weight': 0.9531740695036613,
    'random_state': 42
}

lgbm_gender_params_3 = {
    'n_estimators': 2600,
    'learning_rate': 0.03825929958722602,
    'scale_pos_weight': 0.9531740695036613,
    'random_state': 42
}

cat_gender_best = {
    'iterations': 3000,
    'learning_rate': 0.051418504137325544,
    'max_depth': 4,
    'colsample_bylevel': 0.7594434568905333,
    'l2_leaf_reg': 57.13098420673366,
    'random_strength': 30.89358906898957,
    'bootstrap_type': 'No',
    'border_count': 128,
    'grow_policy': 'Lossguide',
    'od_wait': 1886,
    'leaf_estimation_iterations': 15,
    'eval_metric': 'AUC',
    'scale_pos_weight': 0.9531740695036613,
    'random_state': 42
}

# получим метрики алгоритмов с подобранными параметрами без стекинга
clf1 = LGBMClassifier(**lgbm_gender_best1)
clf1.fit(X_train, y_train)

clf2 = LGBMClassifier(**lgbm_gender_params_2)
clf2.fit(X_train, y_train)

clf3 = LGBMClassifier(**lgbm_gender_params_3).fit(X_train, y_train)
clf3.fit(X_train, y_train)

clf4 = CatBoostClassifier(**cat_gender_best, cat_features=cat_features)
clf4.fit(X_train, y_train, verbose=False)

models_check = [clf1, clf2, clf3, clf4]

metrics = pd.DataFrame()
for model in models_check:
    y_pred = model.predict(X_test)
    y_score = model.predict_proba(X_test)

    metrics = metrics.append(
        get_metrics_classification(y_test,
                                   y_pred,
                                   y_score,
                                   name=f'{model.__class__.__name__}'))

metrics.reset_index(drop=True, inplace=True)

metrics

# LGBM Optuna 1

# кросс-валидация модели и получение предсказаний на фолдах
meta_X = pd.DataFrame()
meta_X_test = pd.DataFrame()

pred_val = []
pred_score_val = []

cv = StratifiedKFold(n_splits=N_FOLDS)

for fold, (train_idx, test_idx) in enumerate(cv.split(X_train, y_train)):
    X_train_, X_val = X_train.iloc[train_idx], X_train.iloc[test_idx]
    y_train_, y_val = y_train.iloc[train_idx], y_train.iloc[test_idx]

    lgbm_1 = LGBMClassifier(**lgbm_gender_best1)
    lgbm_1.fit(X_train_,
               y_train_,
               eval_set=[(X_val, y_val)],
               eval_metric="auc",
               early_stopping_rounds=100,
               verbose=0)

    y_pred_val = lgbm_1.predict(X_val)
    y_score_val = lgbm_1.predict_proba(X_val)

    print("Fold:", fold + 1,
          "ROC-AUC SCORE %.3f" % m.roc_auc_score(y_val, y_score_val[:, 1]))
    print("---")

    # holdout list
    pred_val.append(y_pred_val)
    pred_score_val.append(y_score_val)

# обучаем модель для предсказания на test
lgbm_1.fit(X_train, y_train)

# получим метаданные для обучения и теста мета-модели 
meta_X['lgbm_01'] = np.concatenate(pred_score_val)[:, 1]
meta_X_test['lgbm_01'] = lgbm_1.predict_proba(X_test)[:, 1]

# проверим не переобучилась ли наша модель
check_overfitting_classification(lgbm_1, X_train, y_train, X_test, y_test)

# LGBM Optuna 2

# кросс-валидация модели и получение предсказаний на фолдах
pred_val = []
pred_score_val = []

cv = StratifiedKFold(n_splits=N_FOLDS)

for fold, (train_idx, test_idx) in enumerate(cv.split(X_train, y_train)):
    X_train_, X_val = X_train.iloc[train_idx], X_train.iloc[test_idx]
    y_train_, y_val = y_train.iloc[train_idx], y_train.iloc[test_idx]

    lgbm_2 = LGBMClassifier(**lgbm_gender_params_2)
    lgbm_2.fit(X_train_,
               y_train_,
               eval_set=[(X_val, y_val)],
               eval_metric="auc",
               early_stopping_rounds=100,
               verbose=0)

    y_pred_val = lgbm_2.predict(X_val)
    y_score_val = lgbm_2.predict_proba(X_val)

    print("Fold:", fold + 1,
          "ROC-AUC SCORE %.3f" % m.roc_auc_score(y_val, y_score_val[:, 1]))
    print("---")

    # holdout list
    pred_val.append(y_pred_val)
    pred_score_val.append(y_score_val)

# обучаем модель для предсказания на test
lgbm_2.fit(X_train, y_train)

# получим метаданные для обучения и теста мета-модели 
meta_X['lgbm_02'] = np.concatenate(pred_score_val)[:, 1]
meta_X_test['lgbm_02'] = lgbm_2.predict_proba(X_test)[:, 1]

# проверим не переобучилась ли наша модель
check_overfitting_classification(lgbm_2, X_train, y_train, X_test, y_test)

# LGBM Optuna 3

# кросс-валидация модели и получение предсказаний на фолдах
pred_val = []
pred_score_val = []

cv = StratifiedKFold(n_splits=N_FOLDS)

for fold, (train_idx, test_idx) in enumerate(cv.split(X_train, y_train)):
    X_train_, X_val = X_train.iloc[train_idx], X_train.iloc[test_idx]
    y_train_, y_val = y_train.iloc[train_idx], y_train.iloc[test_idx]

    lgbm_3 = LGBMClassifier(**lgbm_gender_params_3)
    lgbm_3.fit(X_train_,
               y_train_,
               eval_set=[(X_val, y_val)],
               eval_metric="auc",
               early_stopping_rounds=100,
               verbose=0)

    y_pred_val = lgbm_3.predict(X_val)
    y_score_val = lgbm_3.predict_proba(X_val)

    print("Fold:", fold + 1,
          "ROC-AUC SCORE %.3f" % m.roc_auc_score(y_val, y_score_val[:, 1]))
    print("---")

    # holdout list
    pred_val.append(y_pred_val)
    pred_score_val.append(y_score_val)

# обучаем модель для предсказания на test
lgbm_3.fit(X_train, y_train)

# получим метаданные для обучения и теста мета-модели 
meta_X['lgbm_03'] = np.concatenate(pred_score_val)[:, 1]
meta_X_test['lgbm_03'] = lgbm_3.predict_proba(X_test)[:, 1]

# проверим не переобучилась ли наша модель
check_overfitting_classification(lgbm_3, X_train, y_train, X_test, y_test)

# Catboost Optuna

# кросс-валидация модели и получение предсказаний на фолдах
pred_val = []
pred_score_val = []

cv = StratifiedKFold(n_splits=N_FOLDS)

for fold, (train_idx, test_idx) in enumerate(cv.split(X_train, y_train)):
    X_train_, X_val = X_train.iloc[train_idx], X_train.iloc[test_idx]
    y_train_, y_val = y_train.iloc[train_idx], y_train.iloc[test_idx]

    cat_1 = CatBoostClassifier(cat_features=cat_features, **cat_gender_best)
    cat_1.fit(X_train_,
              y_train_,
              eval_set=[(X_val, y_val)],
              early_stopping_rounds=100,
              verbose=False)

    y_pred_val = cat_1.predict(X_val)
    y_score_val = cat_1.predict_proba(X_val)

    print("Fold:", fold + 1,
          "ROC-AUC SCORE %.3f" % m.roc_auc_score(y_val, y_score_val[:, 1]))
    print("---")

    # holdout list
    pred_val.append(y_pred_val)
    pred_score_val.append(y_score_val)

# обучаем модель для предсказания на test
cat_1.fit(X_train, y_train, verbose=False)

# получим метаданные для обучения и теста мета-модели 
meta_X['cat_01'] = np.concatenate(pred_score_val)[:, 1]
meta_X_test['cat_01'] = cat_1.predict_proba(X_test)[:, 1]

# проверим не переобучилась ли наша модель
check_overfitting_classification(cat_1, X_train, y_train, X_test, y_test)

meta_X.to_csv('submissions/meta_X.csv', index=False)
meta_X_test.to_csv('submissions/meta_X_test.csv', index=False)

# Final meta model

В качестве мета-модели будем использовать бейзлайн логистической регрессии

meta_model = LogisticRegression(random_state=RAND)
meta_model.fit(meta_X, y_train)

# проверим не переобучилась ли наша модель
check_overfitting_classification(meta_model, meta_X, y_train, meta_X_test,
                                 y_test)

# получим итоговые метрики для мета-модели
y_pred_test = meta_model.predict(meta_X_test)
y_score_test = meta_model.predict_proba(meta_X_test)

y_pred_train = meta_model.predict_proba(meta_X)

metrics = metrics.append(
    get_metrics_classification(y_test,
                               y_pred_test,
                               y_score_test,
                               name='Stacking_hand'))
metrics

С помощью стекинга удалось улучшить метрики и снизить переобучение
